fixed admin filter in recent actions + fixed error message when exceeding maximum pinned chats number + version bump

+++ app/src/main/java/org/thunderdog/challegram/ui/MessagesController.java
@@ -7888,7 +7888,7 @@ public class MessagesController extends ViewController<MessagesController.Argume
 
                 case R.id.user: {
                   if (item.isSelected() && userIds1 != null) {
-                    userIds1.append(item.getIntValue());
+                    userIds1.append(item.getLongValue());
                   }
                   break;
                 }
